+++ blocklets/ai-studio/src/components/agent-models/model-select-lite.tsx
@@ -1,3 +1,4 @@
+import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import { ImageModelInfo, ModelBasedAssistantYjs, TextModelInfo } from '@blocklet/ai-runtime/types';
 import { Icon } from '@iconify-icon/react';
 import CheckIcon from '@iconify-icons/tabler/check';
@@ -19,11 +20,13 @@ interface InternalModelSelectLiteProps {
 }
 
 function InternalModelSelectLite({ options, value, onChange, onAddMoreModel, ...rest }: InternalModelSelectLiteProps) {
+  const { t } = useLocaleContext();
   const popperState = usePopupState({ variant: 'popper', popupId: 'model-select-lite' });
   const selectedOption = options.find((option) => option.model === value);
+
   return (
     <Box {...rest}>
-      <Button color="inherit" {...bindTrigger(popperState)}>
+      <Button color="inherit" {...bindTrigger(popperState)} data-testid="model-select-lite-trigger">
         {selectedOption?.model && (
           <ModelBrandIcon model={selectedOption.model} url={selectedOption.icon} size="small" sx={{ mr: 0.5 }} />
         )}
@@ -43,11 +46,12 @@ function InternalModelSelectLite({ options, value, onChange, onAddMoreModel, ...
               <ClickAwayListener
                 onClickAway={(e) => (e.target as HTMLElement)?.localName !== 'body' && popperState.close()}>
                 <Stack>
-                  <Box sx={{ maxHeight: 240, p: 1, overflowY: 'auto' }}>
+                  <Box sx={{ maxHeight: 240, p: 1, overflowY: 'auto' }} data-testid="model-select-lite-options">
                     {options.map((option) => {
                       const isSelected = option.model === value;
                       return (
                         <Stack
+                          data-testid={option.model}
                           direction="row"
                           justifyContent="space-between"
                           alignItems="center"
@@ -115,7 +119,7 @@ function InternalModelSelectLite({ options, value, onChange, onAddMoreModel, ...
                         borderColor: 'divider',
                       }}>
                       <Button startIcon={<Icon icon={PlusIcon} />} onClick={onAddMoreModel} sx={{ width: '100%' }}>
-                        Add more model
+                        {t('moreModels')}
                       </Button>
                     </Box>
                   )}