+++ blocklets/ai-studio/src/locales/en.ts
@@ -178,6 +178,7 @@ export default flatten({
       didSpaces: 'DID Spaces',
       appearance: 'Appearance',
       readme: 'README',
+      integrations: 'Integrations',
     },
     selectProject: 'Select Project',
     name: 'Project name',
@@ -920,8 +921,14 @@ export default flatten({
   modelSettings: 'Model Settings',
   modelSettingsPresets: {
     precise: '✨ Precise',
-    balance: '⚖️ Balance',
+    balance: '⚖️ Balanced',
     creative: '🎨 Creative',
     custom: '🖊️ Custom',
   },
+  moreModels: 'More Models',
+  generateObject: 'Generate {object}',
+  clickToGenerateNpmLink: 'Click to generate npm link',
+  modelSettingsTip:
+    'We have some standard settings available for you, and you can make any changes to customize your own parameters.',
+  favorites: 'Favorites',
 });