+++ framework/runtime/src/executor/agent.ts
@@ -1,0 +1,8 @@
+import { Agent } from '../types';
+import { AgentExecutorBase } from './base';
+
+export class AgentExecutor extends AgentExecutorBase<Agent> {
+  override async process() {
+    // ignore
+  }
+}