+++ blocklets/ai-runtime/CHANGELOG.md
@@ -1,3 +1,41 @@
+## 0.4.189 (2024-12-23)
+
+- fix: correct inconsistent header logo display
+
+## 0.4.188 (2024-12-23)
+
+- fix: aigne i18n error
+
+## 0.4.187 (2024-12-21)
+
+- chore: update deps
+
+## 0.4.186 (2024-12-20)
+
+- feat: support llm vision
+
+## 0.4.185 (2024-12-20)
+
+- feat(ai-studio): support for OAuth account to directly access default space
+- feat(ai-studio): view data after sync to DID Space
+
+## 0.4.184 (2024-12-20)
+
+- chore: support copy npm/yarn/pnpm cmd
+- fix: agent type is missing in the generated code
+
+## 0.4.183 (2024-12-20)
+
+- chore(project-models): ui improvements
+
+## 0.4.182 (2024-12-20)
+
+- Support install a project as npm package
+
+## 0.4.177 (2024-12-18)
+
+- fix: data not loading after changing resource type
+
 ## 0.4.176 (2024-12-17)
 
 - chore: update start timeout for blocklet start